+++ services/backend/src/router/admin/validator.js
@@ -1,12 +1,5 @@
-const { query, body } = require("express-validator");
+const { body } = require("express-validator");
 const { isUUID, isIn } = require("validator");
-
-const langValidator = [
-  query("language")
-    .trim()
-    .isIn(["ENGLISH", "FRENCH"])
-    .withMessage("must be 'ENGLISH' or 'FRENCH'"),
-];
 
 const updateUserStatusValidator = [
   body()
@@ -22,6 +15,5 @@ const updateUserStatusValidator = [
 ];
 
 module.exports = {
-  langValidator,
   updateUserStatusValidator,
 };
