+++ services/backend/src/router/connections/connections.js
@@ -3,7 +3,7 @@ const { keycloak } = require("../../auth/keycloak");
 
 const connections = require("../../core/connections/connections");
 
-const { UUIDValidator } = require("./validator");
+const { UUIDValidator } = require("../util/commonValidators");
 
 const connectionsRouter = Router();
