--- services/backend/src/router/geds/validator.js
@@ -1,7 +1,0 @@
-const { param } = require("express-validator");
-
-const UUIDValidator = param("id").trim().isUUID().withMessage("must be a UUID");
-
-module.exports = {
-  UUIDValidator,
-};
