+++ services/backend/src/router/options/docs.yml
@@ -202,3 +202,25 @@
     security:
       - Keycloak: []
     tags: [option]
+
+/api/option/attachmentNames:
+  get:
+    summary: Returns an array of attachment names
+    produces:
+      - application/json
+    parameters:
+      - $ref: "#/definitions/LanguageQuery"
+      - $ref: "#/definitions/TypeQuery"
+      -
+    responses:
+      200:
+        $ref: "#/definitions/IdNameArray"
+      403:
+        $ref: "#/definitions/403"
+      422:
+        $ref: "#/definitions/422"
+      500:
+        $ref: "#/definitions/500"
+    security:
+      - Keycloak: []
+    tags: [option]
