+++ services/backend/src/router/options/validator.js
@@ -1,12 +1,5 @@
-const { query, body } = require("express-validator");
+const { body } = require("express-validator");
 const { isUUID } = require("validator");
-
-const langValidator = [
-  query("language")
-    .trim()
-    .isIn(["ENGLISH", "FRENCH"])
-    .withMessage("must be 'ENGLISH' or 'FRENCH'"),
-];
 
 const deleteManyValidator = [
   body("ids")
@@ -50,7 +43,6 @@ const createSkillValidator = [...createValidator, skillValidator];
 const updateSkillValidator = [...updateValidator, skillValidator];
 
 module.exports = {
-  langValidator,
   deleteManyValidator,
   deleteOneValidator,
   createValidator,
