+++ services/backend/src/router/profile/validator.js
@@ -1,13 +1,14 @@
-const { query, param, body } = require("express-validator");
+const { body } = require("express-validator");
 const { isUUID, isIn, isMobilePhone } = require("validator");
 const moment = require("moment");
 
-const langValidator = query("language")
-  .trim()
-  .isIn(["ENGLISH", "FRENCH"])
-  .withMessage("must be 'ENGLISH' or 'FRENCH'");
-
-const UUIDValidator = param("id").trim().isUUID().withMessage("must be a UUID");
+const updateProfilePhoneNumberBody = ["telephone", "cellphone"];
+const updateProfileNumberBody = ["signupStep"];
+const updateProfileStringArrayBody = ["projects"];
+const updateProfileDateBody = ["actingStartDate", "actingEndDate"];
+const updateProfileOptionalLanguageBody = ["firstLanguage", "secondLanguage"];
+const updateProfileLanguageBody = ["preferredLanguage"];
+const updateProfileBooleanBody = ["interestedInRemote", "exFeeder"];
 
 const updateProfileStringBody = [
   "firstName",
@@ -19,18 +20,6 @@ const updateProfileStringBody = [
   "manager",
   "avatarColor",
 ];
-
-const updateProfilePhoneNumberBody = ["telephone", "cellphone"];
-
-const updateProfileNumberBody = ["signupStep"];
-const updateProfileStringArrayBody = ["projects"];
-const updateProfileDateBody = ["actingStartDate", "actingEndDate"];
-
-const updateProfileOptionalLanguageBody = ["firstLanguage", "secondLanguage"];
-
-const updateProfileLanguageBody = ["preferredLanguage"];
-
-const updateProfileBooleanBody = ["interestedInRemote", "exFeeder"];
 
 const updateProfileUUIDArrayBody = [
   "skills",
@@ -53,8 +42,6 @@ const updateProfileUUIDBody = [
 ];
 
 const updateProfileValidator = [
-  langValidator,
-  UUIDValidator,
   updateProfileStringBody.map((i) =>
     body(i)
       .optional()
@@ -186,7 +173,6 @@ const updateProfileValidator = [
 ];
 
 module.exports = {
-  langValidator,
   UUIDValidator,
   updateProfileValidator,
 };
