--- services/backend/src/router/statistics/validator.js
@@ -1,12 +1,0 @@
-const { query } = require("express-validator");
-
-const langValidator = [
-  query("language")
-    .trim()
-    .isIn(["ENGLISH", "FRENCH"])
-    .withMessage("must be 'ENGLISH' or 'FRENCH'"),
-];
-
-module.exports = {
-  langValidator,
-};
