+++ services/backend/src/router/util/commonValidators.js
@@ -1,4 +1,4 @@
-const { query } = require("express-validator");
+const { param, query } = require("express-validator");
 
 const langValidator = [
   query("language")
@@ -7,6 +7,9 @@ const langValidator = [
     .withMessage("must be 'ENGLISH' or 'FRENCH'"),
 ];
 
+const UUIDValidator = param("id").trim().isUUID().withMessage("must be a UUID");
+
 module.exports = {
   langValidator,
+  UUIDValidator,
 };
