+++ services/frontend-v3/src/components/education/EducationView.jsx
@@ -20,14 +20,14 @@ const EducationView = ({ educationInfo }) => {
   const getUrl = (item) => {
     if (item.attachmentLinks && item.attachmentLinks.length > 0)
       return item.attachmentLinks.map((i) => (
-        <Tag color="rgb(114, 114, 114)" key={i.id}>
-          <LinkOutlined />
-          <a target="_blank" rel="noreferrer" href={i.url}>
-            {i.name}
-          </a>
-        </Tag>
+        <a target="_blank" rel="noreferrer" href={i.url}>
+          <Tag color="rgb(114, 114, 114)" key={i.id}>
+            <LinkOutlined />
+            <span>{i.name}</span>
+          </Tag>
+        </a>
       ));
-    return <></>;
+    return undefined;
   };
 
   const generateEducationItemDescription = (item) => (
