+++ services/frontend-v3/src/i18n/en_CA.json
@@ -108,6 +108,7 @@
   "advanced.search.form.mentorship.skills": "Mentorship Skills",
   "advanced.search.form.name": "Name",
   "advanced.search.form.skills": "Skills",
+  "attachment.placeholder": "Attachment URL - ex. https://www.google.com",
   "back.to.landing": "Back to landing",
   "button.advanced.search": "Advanced search",
   "button.basic.search": "Basic search",
