+++ services/frontend-v3/src/i18n/fr_CA.json
@@ -108,6 +108,7 @@
   "advanced.search.form.mentorship.skills": "Compétences de mentorat",
   "advanced.search.form.name": "Nom",
   "advanced.search.form.skills": "Compétences",
+  "attachment.placeholder": "URL de la pièce jointe - ex. https://www.google.com",
   "back.to.landing": "Retour à la page d'accueil",
   "button.advanced.search": "Recherche avancée",
   "button.basic.search": "Recherche de base",
